# docs/swagger/foedie.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ecommerce Auth API",
    "version": "1.0",
    "contact": {
      "name": "Lovelyo",
      "url": "https://lovelyoyrmia.com",
      "email": "lovelyoyrmia@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/carts": {
      "get": {
        "summary": "Get Cart Products",
        "description": "Use this API to get products in cart",
        "operationId": "OrderService_GetCarts",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetCartUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OrderService"
        ]
      },
      "post": {
        "summary": "Add Cart",
        "description": "Use this API to add products to cart",
        "operationId": "OrderService_AddCart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateCartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateCartParams"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/api/v1/carts/{oid}": {
      "delete": {
        "summary": "Delete Cart Products",
        "description": "Use this API to delete products in cart",
        "operationId": "OrderService_DeleteCartProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteCartProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pid",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
...
  },
  "definitions": {
    "pbCartProduct": {
      "type": "object",
      "properties": {
        "pid": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stock": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "string",
          "format": "int64"
        },
        "category": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbCreateCartParams": {
      "type": "object",
      "properties": {
        "pid": {
          "type": "string"
        },
        "uid": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbCreateCartResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "pbDeleteCartProductResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "pbGetCartUserResponse": {
      "type": "object",
      "properties": {
        "oid": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbCartProduct"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}